% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_removal.R
\name{batch_removal}
\alias{batch_removal}
\title{batch_removal with sva's ComBat() function}
\usage{
batch_removal(expmatrix, pheno)
}
\arguments{
\item{expmatrix}{An input matrix with gene names as row names and sample names as column names.}

\item{pheno}{Phenotype table with batch information.}
}
\value{
Returns batch corrected expression matrix.
}
\description{
Removes batch effects in combined datasets using sva's ComBat() function.
}
\examples{
pheno <- data.frame(sample=c(1:16), treatment=sub("(.*)(_[0-9])", "\\\\1", colnames(expmatrix)),
                   batch=ifelse(grepl("Ctrl", colnames(expmatrix)) == TRUE, "1",
                   ifelse(grepl("ActLPS", colnames(expmatrix)) == TRUE, "1", "2")),
                   row.names = colnames(expmatrix))
batch_removal(expmatrix, pheno)
}

