% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlationPlot}
\alias{correlationPlot}
\title{Calculates and Produces Correlation Plot of Genes}
\usage{
correlationPlot(expressionData)
}
\arguments{
\item{expressionData}{A dataframe of gene expression data, with genes as
rows and samples as columns.}
}
\value{
Returns a plot of pairwise correlations between genes
}
\description{
A function that calculates the pairwise-correlation of genes, given a matrix
of gene expression in different samples, helping to identify correlated
genes. The function uses the \code{corrplot} package to create the plot.
}
\examples{
# Using OVExpression dataset in package
correlationPlot(expressionData = OVExpression)

}
\references{
Wei T, Simko V (2021). R package 'corrplot': Visualization of a Correlation
Matrix. (Version 0.92), \url{https://github.com/taiyun/corrplot}.
}
