% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataOutput}
\alias{dataOutput}
\title{Example data output from WGCNA}
\format{A data frame with 10000 rows and 21 variables. Rownames show sample names.
\describe{
  \item{input}{Gene name}
  \item{ENTREZID}{ENTREZ ID}
  \item{ENSEMBL}{ENSEMBL ID}
  \item{moduleColors}{Module membership color}
  \item{MM.blue}{Module membership value blue module}
  \item{MM.brown}{Module membership value brown module}
  \item{MM.pink}{Module membership value pink module}
  \item{MM.red}{Module membership value red module}
  \item{MM.green}{Module membership value green module}
  \item{MM.magenta}{Module membership value magenta module}
  \item{MM.black}{Module membership value black module}
  \item{MM.yellow}{Module membership value yellow module}
  \item{MM.purple}{Module membership value purple module}
  \item{MM.turquoise}{Module membership value turquoise module}
  \item{MM.grey}{Module membership value grey module}
  \item{cor.Ctrl}{Correlation with control samples}
  \item{cor.TolLPS}{Correlation with TolLPS samples}
  \item{cor.TolS100A8}{Correlation with TolS100A8 samples}
  \item{cor.ActLPS}{Correlation with ActLPS samples}
  \item{cor.Tol}{Correlation with Tol samples}
  \item{cor.ExPhenotype}{Correlation with ExPhenotype samples}
}}
\source{
WGCNA example data output
}
\usage{
data(dataOutput)
}
\description{
A dataset containing example output from WGCNA analysis.
}
\keyword{datasets}

