% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEgenes_pw}
\alias{DEgenes_pw}
\title{Output data frame from diff_limma_pairwise()}
\format{A data frame with 9956 rows and 8 variables. Rownames show gene names.
\describe{
  \item{logFC}{log fold change}
  \item{CI.L}{left confidence interval}
  \item{CI.R}{right confidence interval}
  \item{AveExpr}{Average expression}
  \item{t}{t statstics}
  \item{P.Value}{p-value}
  \item{adj.P.Val}{adjusted p-value}
  \item{B}{B statistics}
}}
\source{
Output data frame from diff_limma_pairwise() (R package "exprAnalysis")
}
\usage{
DEgenes_pw
}
\description{
A dataset containing output data frame from diff_limma_pairwise().
}
\keyword{datasets}

