% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_limma.R
\name{diff_limma_all}
\alias{diff_limma_all}
\title{diff_limma_all (from limma functions).}
\usage{
diff_limma_all(expmatrix, design, groupcomparisons, p.value = 0.05,
  lfc = 1.5, projectfolder = getwd())
}
\arguments{
\item{expmatrix}{An input matrix with gene names as row names and sample names as column names.}

\item{design}{Binary design matrix of samples' treatment groups.}

\item{groupcomparisons}{Group comparisons for differential expression analysis.}

\item{p.value}{Significance threshold. Defaults to 0.05}

\item{lfc}{Log fold change threshold. Defaults to 1.5}

\item{projectfolder}{File path where to save the output to. Defaults to working directory. Here, it saves the output to a subfolder called "Diff_limma".}
}
\value{
Differential expression output table for all group comparisons - outputs information for all genes, including non-significant ones.
}
\description{
This calculates the differentially expressed genes among all group comparisons.
}
\examples{
design <- data.frame(Ctrl = c(rep(1, 4), rep(0,12)), TolLPS = c(rep(0, 4), rep(1, 4),
             rep(0, 8)), TollMRP8 = c(rep(0, 8), rep(1, 4), rep(0, 4)), ActLPS = c(rep(0, 12),
             rep(1, 4)), row.names = colnames(expmatrix))
groupcomparisons=c("TolLPS-Ctrl", "TollMRP8-Ctrl", "ActLPS-Ctrl")
diff_limma_all(expmatrix, design, groupcomparisons)
}

