% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_limma_pw.R
\name{diff_limma_pw_unfiltered}
\alias{diff_limma_pw_unfiltered}
\title{diff_limma_pw_unfiltered (from limma functions).}
\usage{
diff_limma_pw_unfiltered(expmatrix, design, comparison,
  projectfolder = getwd())
}
\arguments{
\item{expmatrix}{An input matrix with gene names as row names and sample names as column names.}

\item{design}{Binary design matrix of samples' treatment groups.}

\item{comparison}{Single comparisons for differential expression analysis.}

\item{projectfolder}{File path where to save the output to. Defaults to working directory. Here, it saves the output to a subfolder called "Diff_limma".}
}
\value{
Unfiltered Differential expression output table for the given comparison. The table and Venn diagram are saved to files in the folder "diff_limma" and the table is also returned by the function.
}
\description{
This calculates the differentially expressed genes among pairwise group comparisons (one comparison at a time) and outputs results for all genes (including non-significant genes).
}
\examples{
design <- data.frame(Ctrl = c(rep(1, 4), rep(0,12)), TolLPS = c(rep(0, 4), rep(1, 4),
             rep(0, 8)), TollMRP8 = c(rep(0, 8), rep(1, 4), rep(0, 4)), ActLPS = c(rep(0, 12),
             rep(1, 4)), row.names = colnames(expmatrix))
comparison="TolLPS-Ctrl"
diff_limma_pw_unfiltered(expmatrix, design, comparison)
}

