% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprAnalysis.R
\docType{package}
\name{exprAnalysis}
\alias{exprAnalysis}
\alias{exprAnalysis-package}
\title{exprAnalysis: A package for analysing RNA-seq and expression array data.}
\description{
exprAnalysis: A package for analysing RNA-seq and expression array data.
}
\section{exprAnalysis functions}{

The exprAnalysis functions ...
}

