% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{exprNormalization}
\alias{exprNormalization}
\title{Normalizes Gene Expression Data}
\usage{
exprNormalization(expressionData, method = "log")
}
\arguments{
\item{expressionData}{A dataframe of numeric gene expression data, with genes as
rows and samples as columns.}

\item{method}{A parameter to specify normalization method. The options are:
\itemize{
\item "log" - log2 transformation (default)
\item "total" - total count normalization
\item "standard" - z-score standardization
}}
}
\value{
Returns a dataframe of of normalized gene expression data from
expressionData.
}
\description{
A function that calculates normalizes gene expression data, with three
methods available: total count normalization, log2 transformation, and
z-score standardization. The function uses the \code{dplyr} package to
manipulate the data.
}
\examples{
# Using OVExpression dataset in package

# Example 1:
# Using default normalization method "log"
logExpr <- exprNormalization(OVExpression)
head(logExpr[ , 1:5])


# Example 2:
# Using total count normalization
normalizedExpr <- exprNormalization(OVExpression, method = "total")
head(normalizedExpr[ , 1:5])

# sum of each column/sample is 1
colSums(normalizedExpr)


# Example 3:
# Using standardization
standardizedExpr <- exprNormalization(OVExpression, method = "standard")
head(standardizedExpr[ , 1:5])

# mean for each column/sample is 0
round(colMeans(standardizedExpr))
# standard deviation for each column/sample is 1
apply(standardizedExpr,2,sd)

}
\references{
Wickham H, François R, Henry L, Müller K (2022). dplyr: A Grammar of Data
Manipulation. \url{https://dplyr.tidyverse.org},
\url{https://github.com/tidyverse/dplyr}.
}
