% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{exprPCA}
\alias{exprPCA}
\title{Create a PCA Plot}
\usage{
exprPCA(expressionData, sampleData)
}
\arguments{
\item{expressionData}{A dataframe of numeric gene expression data, with genes as
rows and samples as columns.}

\item{sampleData}{A dataframe of sample information, with samples as rownames,
and a column called "type" identifying if the sample is case or control.}
}
\value{
Produces a PCA Plot, with samples coloured according to their type
}
\description{
A function that creates a PCA plot based on the expression data.
}
\examples{
# Example
exprPCA(OVExpression, OVSample)

}
\references{
Kassambara A, Mundt F (2020). \emph{factoextra: Extract and Visualize the Results of Multivariate
Data Analyses}. R package version 1.0.7, \url{https://CRAN.R-project.org/package=factoextra}.

Sebastien Le, Julie Josse, Francois Husson (2008). FactoMineR: An R Package
for Multivariate Analysis. Journal of Statistical Software, 25(1), 1-18.
}
