% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heatmaps}
\alias{heatmaps}
\title{heatmaps (using heatmaps.2 of "gplots").}
\usage{
heatmaps(expmatrix, method_dist = "canberra", method_hclust = "ward.D2",
  samplecols, main = "Heatmap of genes")
}
\arguments{
\item{expmatrix}{An input matrix with gene names as row names and sample names as column names.}

\item{method_dist}{the distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given.}

\item{method_hclust}{the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}
}
\value{
Returns a heatmap plot.
}
\description{
This calculates the differentially expressed genes among pairwise group comparisons (one comparison at a time).
}
\examples{
heatmaps(expmatrix[1:100,], samplecols = rep(c("#E41A1C", "#377EB8", "#4DAF4A", "#984EA3"), each=4))
}

