% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggDisease.R
\name{keggDiseaseGenes}
\alias{keggDiseaseGenes}
\title{Gets Genes Associated with KEGG Disease Entry}
\usage{
keggDiseaseGenes(entryID)
}
\arguments{
\item{entryID}{A string of the KEGG disease entry ID}
}
\value{
Returns a character vector containing genes from the inputted entry
}
\description{
A function that returns the genes associated with a KEGG disease entry, using
the function keggGET from the KEGGREST package.
}
\examples{
# Example:
# Getting genes for KEGG entry for Ovarian cancer (H00027)
keggDiseaseGenes(entryID = "H00027")

}
\references{
Tenenbaum D, Maintainer B (2022). \emph{KEGGREST: Client-side REST access to the
Kyoto Encyclopedia of Genes and Genomes (KEGG)}. R package version 1.36.3.

Wickham H (2022). \emph{stringr: Simple, Consistent Wrappers for Common String
Operations}. R packageversion 1.5.0, \url{https://CRAN.R-project.org/package=stringr}.
}
