% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load data from files}
\usage{
loadData(exprFilePath, sampleFilePath, sep = "tab")
}
\arguments{
\item{exprFilePath}{A string of the path to the file containing gene
expression data. The file should be a matrix with the first row as sample
names, the first column as gene names, and expression values in the rest
of the matrix.}

\item{sampleFilePath}{A string of the path to the file containing sample
information. The file should have the first column as sample names, and
the second column as their type.}

\item{sep}{A parameter to specify the the separator in the file. The options
are:
\itemize{
\item "tab" - for whitespace/tab-separated values (default)
\item "comma" - for comma-separated values
\item "semicolon" - for semicolon-separated values
}}
}
\value{
Returns a list of two dataframes:
\itemize{
\item expressionData - dataframe of expression values, rows are genes, columns are samples

\item sampleData - dataframe of sample information, rows are samples, one column "type" with the type of sample
}
}
\description{
A function prepares data in the format required for other functions in the
expressionAnalysis package, given the path to the files.
}
\examples{
# Access data made available with this package
OVExpresssionPath <- system.file("extdata", "OVExpression.CSV", package = "expressionAnalysis")
OVSamplePath <- system.file("extdata", "OVSample.CSV", package = "expressionAnalysis")

# Load data with loadData function
loadedData <- loadData(OVExpresssionPath, OVSamplePath, sep = "comma")

# Access the dataframes
loadedData$expressionData    # dataframe for expression data
loadedData$sampleData        # dataframe for sample information

}
