% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEs}
\alias{MEs}
\title{Module Eigengenes example data}
\format{A data frame with 16 rows and 11 variables. Rownames show sample names.
\describe{
  \item{MEblue}{Module eigengenes of blue module}
  \item{MEbrown}{Module eigengenes of brown module}
  \item{MEpink}{TModule eigengenes of pink module}
  \item{MEred}{TModule eigengenes of red module}
  \item{MEgreen}{TModule eigengenes of green module}
  \item{MEmagenta}{TModule eigengenes of magenta module}
  \item{MEblack}{TModule eigengenes of black module}
  \item{MEyellow}{TModule eigengenes of yellow module}
  \item{MEpurple}{TModule eigengenes of purple module}
  \item{MEturquoise}{TModule eigengenes of turquoise module}
  \item{MEgrey}{TModule eigengenes of grey module}
}}
\source{
WGCNA example data run.
}
\usage{
data(MEs)
}
\description{
A dataset containing module eigengenes example data.
}
\keyword{datasets}

