% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_eset.R
\name{normalise_eset}
\alias{normalise_eset}
\title{normalise_eset (using beadarray).}
\usage{
normalise_eset(eset, method_norm = "quantile", transform = "none")
}
\arguments{
\item{eset}{expression set}

\item{method_norm}{Normalisation method. See ?beadarray::readBeadSummaryData for details}

\item{transform}{Transformation method. See ?beadarray::readBeadSummaryData for details}
}
\value{
Returns (normalized) expression set object.
}
\description{
Reads in Illumina expression data.
}

