% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OVExpression}
\alias{OVExpression}
\title{Gene Expression from an Ovarian Cancer Experiment}
\format{
A matrix with rows as genes, and columns as samples. The values are
gene expressions for corresponding genes in corresponding samples.
}
\source{
NCBI Gene Expression Omnibus (Accession
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE14407}{GSE14407})
}
\usage{
OVExpression
}
\description{
Data from ovarian cancer gene expression profiling of 24 samples (12 normal
human ovaries and 12 ovarian cancer epithelial cells). Only a subset of genes
(10 genes) were kept for the example expression data in this package.
}
\examples{
\dontrun{
 OVExpression
}
}
\references{
Bowen NJ, Walker LD, Matyunina LV, Logani S et al. Gene expression profiling
supports the hypothesis that human ovarian surface epithelia are multipotent
and capable of serving as ovarian cancer initiating cells. BMC Med Genomics
2009 Dec 29;2:71. PMID: \href{https://pubmed.ncbi.nlm.nih.gov/20040092/}{20040092}
}
\keyword{datasets}
