% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaplot.R
\name{pca_plot}
\alias{pca_plot}
\title{pca_plot (from pcaGoPromoter functions).}
\usage{
pca_plot(expmatrix, groups, PCs = c(1, 2), main = "PCA plot")
}
\arguments{
\item{expmatrix}{An input matrix with gene names as row names and sample names as column names.}

\item{groups}{Define sample groups.}

\item{PCs}{Principle components to plot (pick two). Defaults to PCs 1 and 2.}

\item{main}{Main title. Defaults to "PCA plot".}
}
\value{
A PCA plot.
}
\description{
This is a collection of functions from the package "pcaGoPromoter".
It calculates the principle components and plots them.
}
\examples{
 groups <- as.factor(c(rep("control",4), rep("TolLPS",4), rep("TolS100A8",4), rep("ActLPS",4)))
 pca_plot(expmatrix, groups)
}

