% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaGoPromoter.R
\name{pca_plot_enrich}
\alias{pca_plot_enrich}
\title{pca_plot_enrich (from pcaGoPromoter functions).}
\usage{
pca_plot_enrich(expmatrix, groups, PCs = c(1, 2), main = "PCA plot",
  percentage = 2.5, inputType = "geneSymbol", org = "Hs",
  projectfolder = getwd())
}
\arguments{
\item{expmatrix}{An input matrix with gene names as row names and sample names as column names.}

\item{groups}{Define sample groups.}

\item{PCs}{Principle components to plot (pick two). Defaults to PCs 1 and 2.}

\item{main}{Main title of PCA plot. Defaults to "PCA plot".}

\item{percentage}{Percentage of genes with top loadings to use for TF and GO enrichment analysis. Defauls to 2.5.}

\item{inputType}{a character vector description the input type. Must be Affymetrix chip type, "geneSymbol" or "entrezID". The following Affymetrix chip type are supported: hgu133plus2, mouse4302, rat2302, hugene10st and mogene10st. Default is Affymetrix chip type "hgu133plus2".}

\item{org}{a character vector with the organism. Can be "Hs", "Mm" or "Rn". Only needed if inputType is "geneSymbol" or "entrezID". See details. Default is "Hs".}

\item{projectfolder}{File path where to save the tables to. Defaults to working directory. Here, it saves the output to a subfolder called "pcaGoPromoter".}
}
\value{
A PCA plot and tables of enriched TFs and GO terms.
}
\description{
This is a collection of functions from the package "pcaGoPromoter".
It calculates the principle components and plots them.
It also calculates TF and GO term enrichments.
}
\details{
Both, PC plot (as .pdf) and tables (as tab delimited .txt) of enriched TFs and GO terms are saved by default to the working directory or to the given path.
}
\examples{
 expmatrix <- read.table(system.file("extdata", "Random_exprMatrix.txt", package = "ExpressionAnalysis"),
               header = TRUE, sep = "\\t")
 groups <- as.factor(c(rep("control",4), rep("TolLPS",4), rep("TolS100A8",4), rep("ActLPS",4)))
 pca_plot_enrich(expmatrix, groups)
}

