% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_plots.R
\name{quality_control_plots}
\alias{quality_control_plots}
\title{quality_control_plots (using arrayQualityMetrics).}
\usage{
quality_control_plots(eset, projectfolder = getwd(),
  groupColumn = "Sample_Group")
}
\arguments{
\item{eset}{expression set object.}

\item{projectfolder}{Projectfolder. Creates a subfolder "QC" here and saves output to it.}

\item{groupColumn}{Column name denoting treatment information}
}
\value{
Produces quality control plots.
}
\description{
Produces quality control plots.
}

