% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Illumina.R
\name{read_Illumina}
\alias{read_Illumina}
\title{read_Illumina (using beadarray).}
\usage{
read_Illumina(dataFile, qcFile, sampleSheet, expressionchipType,
  ProbeID = "PROBE_ID", controlID = "ProbeID", method_norm = "none",
  transform = "log2")
}
\arguments{
\item{dataFile}{SampleProbeProfile.}

\item{qcFile}{ControlProbeProfile.}

\item{sampleSheet}{samplesheet.}

\item{expressionchipType}{Expression chip name.}

\item{ProbeID}{Column name of probe ID in SampleProbeProfile.}

\item{controlID}{Column name of probe ID in ControlProbeProfile}

\item{method_norm}{Normalisation method. See ?beadarray::readBeadSummaryData for details.}

\item{transform}{Transformation method. See ?beadarray::readBeadSummaryData for details.}
}
\value{
Returns expression set object.
}
\description{
Reads in Illumina expression data.
}
\examples{
read_Illumina(dataFile, qcFile, sampleSheet, expressionchipType)
}

