% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExpressionAnalysis.R
\name{runExpressionAnalysis}
\alias{runExpressionAnalysis}
\title{Launch Shiny App for expressionAnalysis}
\usage{
runExpressionAnalysis()
}
\value{
No return value but opens up a Shiny page.
}
\description{
add description
}
\examples{
\dontrun{

expressionAnalysis::runExpressionAnalysis()
}

}
\references{
Chang W, Cheng J, Allaire J, Sievert C, Schloerke B, Xie Y, Allen J, McPherson J, Dipert A,
Borges B (2022). \emph{shiny: Web Application Framework for R}. R package version 1.7.3,
\url{https://CRAN.R-project.org/package=shiny}.
}
