% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_networks.R
\name{TF_networks}
\alias{TF_networks}
\title{TF_networks.}
\usage{
TF_networks(expmatrix, nodeAnno, GeneName = "row.names",
  projectfolder = getwd(), organism = "human", outPrefix = "outPrefix")
}
\arguments{
\item{expmatrix}{An input matrix with gene names as row names and sample names as column names.}

\item{nodeAnno}{A data frame with node annotation, e.g. logFC and differential expression (e.g. limma output). Rownames have to correspond to rownames of expmatrix!}

\item{GeneName}{Column name with Gene Symbol to be annotated to TF name.}

\item{projectfolder}{File path where to save the output to. Defaults to working directory. Here, it saves the output to a subfolder called "Networks".}

\item{organism}{Organism. Can be human or mouse. For human, the longer published list of TFs is used; for mouse the shorter list provided by Bonn (for which I don't have any more info on where it comes from).}

\item{outPrefix}{Prefix added to output name.}
}
\value{
A .expression output matrix of gene expression of transcription factors in dataset used as Biolayout3D input.
And based on these TFs, it also produces a .txt with node annotations of e.g. logFC and differential expression (e.g. limma output)
}
\description{
Ceates the input files for TF network analysis using Biolayout3D and Cytoscape.
}
\details{
The .expression matrix of TF expression data will have to be opened with Biolayout3D: Set Minimum Correlation and Correlation metric (by default 0.7 and Pearson). Then choose a suitable correlation coefficient (Graph Degree Distribution should be close to linear). Save the resulting network as a TGF file.
Open the TGF file in Cytoscape: Open network from file. Then got to Advanced Options: untick "Use first column names" and add " to "Other:". Now set Column 2 as Source and Column 5 as Target.
Then open the node annotation table that was also saved by this function: Open table from file, import data as Node Table Columns. YOu can then customize the look of your network.
}
\examples{
TF_networks(expmatrix, nodeAnno=Allgenes_limma_pw)
}

