% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TFs}
\alias{TFs}
\title{Transcription factors from mouse and human}
\format{A data frame with 929 rows and 3 variables. Rownames show gene names.
\describe{
  \item{Mouse}{Mouse TFs}
  \item{Human}{Human TFs}
  \item{Merged_Taxa}{Type of TF}
}}
\source{
Bonner list used for TF networks
}
\usage{
data(TFs)
}
\description{
A dataset containing transcription factors from mouse and human (from Bonn)
}
\keyword{datasets}

