% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TFs_paper}
\alias{TFs_paper}
\title{Transcription factors from human (published list)}
\format{A data frame with 1374 rows and 3 variables. Rownames show gene names.
\describe{
  \item{Human}{Human TFs}
  \item{NodeType}{0 = TF}
  \item{GeneType}{gene type}
}}
\source{
Automated Identification of Core Regulatory Genes in Human Gene Regulatory Networks. Vipin Narang et al. 2015. PLOS Comp Biol.
}
\usage{
data(TFs_paper)
}
\description{
A dataset containing transcription factors from human (from paper: Automated Identification of Core Regulatory Genes in Human Gene Regulatory Networks. Vipin Narang et al. 2015. PLOS Comp Biol.)
}
\keyword{datasets}

