% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_vs_mean.R
\name{var_vs_mean}
\alias{var_vs_mean}
\title{var_vs_mean}
\usage{
var_vs_mean(expmatrix)
}
\arguments{
\item{expmatrix}{An input matrix with gene names as row names and sample names as column names.}
}
\value{
Returns a plot of mean against variance of gene expression across samples and cor.test output of the linear correlation.
}
\description{
This function plots mean against variance of gene expression across samples and calculates the linear correlation.
}

