% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_heatmaps.R
\name{wgcna_heatmaps}
\alias{wgcna_heatmaps}
\title{wgcna_heatmaps (from WGCNA functions).}
\usage{
wgcna_heatmaps(MEs, datExpr, datTraits, projectfolder = getwd())
}
\arguments{
\item{MEs}{Module eigengenes.}

\item{datExpr}{Tranposed input matrix with gene names as column names and sample names as row names.}

\item{datTraits}{Phenotype data.}

\item{projectfolder}{File path where to save the output to. Defaults to working directory. Here, it saves the output to a subfolder called "WGCNA".}
}
\value{
Writes Module_trait_correlations.txt and Module_CorPval.txt Plots module-trait correlation heatmap.
}
\description{
This function implements module-trait correlation heatmap plotting with wgcna.
}
\examples{
wgcna_heatmaps(MEs, datExpr, datTraits)
}

