% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_modulememberships.R
\name{wgcna_modulememberships}
\alias{wgcna_modulememberships}
\title{wgcna_modulememberships (from WGCNA functions).}
\usage{
wgcna_modulememberships(datExpr, datTraits, MEs, moduleColors,
  projectfolder = getwd())
}
\arguments{
\item{datExpr}{Tranposed input matrix with gene names as column names and sample names as row names.}

\item{datTraits}{Phenotype data.}

\item{MEs}{Module eigengenes.}

\item{moduleColors}{Module colors.}

\item{projectfolder}{File path where to save the output to. Defaults to working directory. Here, it saves the output to a subfolder called "WGCNA".}
}
\value{
Writes geneModuleMembership.txt and geneModuleMembershipPvalue.txt. Saves Intramodular analysis plots to pdf. Returns datKME.
}
\description{
This function implements module membership analysis with wgcna.
}

