% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_plotDendroAndColors.R
\name{wgcna_plotDendroAndColors}
\alias{wgcna_plotDendroAndColors}
\title{wgcna_plotDendroAndColors (from WGCNA functions).}
\usage{
wgcna_plotDendroAndColors(datExpr, datTraits, net)
}
\arguments{
\item{datExpr}{Tranposed input matrix with gene names as column names and sample names as row names.}

\item{datTraits}{Phenotype data.}

\item{net}{Output from blockwiseModules().}
}
\value{
Plots module-trait correlation dendrogram
}
\description{
This function implements module-trait correlation dendrogram plotting with wgcna.
}

