% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_removal.R
\name{batch_removal}
\alias{batch_removal}
\title{batch_removal with sva's ComBat() function}
\usage{
batch_removal(expmatrix, pheno)
}
\arguments{
\item{expmatrix}{An input matrix with gene names as row names and sample names as column names.}

\item{pheno}{Phenotype table with batch information.}
}
\value{
Returns batch corrected expression matrix.
}
\description{
Removes batch effects in combined datasets using sva's ComBat() function.
}
\examples{
test_expmatrix <- matrix(rnorm(100*30),nrow=100,ncol=8,
                         dimnames=list(c(rownames(expmatrix[1:100,1:8])),
                         c(colnames(expmatrix[1:100,1:8]))))
pheno <- data.frame(batch=as.factor(rep(c(1,2), 4)), row.names = colnames(test_expmatrix))
test_expmatrix <- batch_removal(expmatrix=test_expmatrix, pheno)
}

