% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlationPlot}
\alias{correlationPlot}
\title{Calculates and Produces Correlation Plot of Genes}
\usage{
correlationPlot(expressionData, method = "spearman")
}
\arguments{
\item{expressionData}{A dataframe of numeric gene expression data, with genes as
rows and samples as columns.}

\item{method}{A parameter to specify correlation coefficient to compute.
The options are:
\itemize{
\item "spearman" (default)
\item "pearson"
\item "kendall"
}}
}
\value{
Returns a plot of pairwise correlations between genes
}
\description{
A function that calculates the pairwise-correlation of genes, given a matrix
of gene expression in different samples, helping to identify correlated
genes. The function uses the \code{corrplot} package to create the plot.
}
\examples{
# Using OVExpression dataset in package
# Example 1:
# Using default correlation coefficient "spearman"
correlationPlot(expressionData = OVExpression)

# Example 2:
# Using pearson as method
correlationPlot(expressionData = OVExpression, method = "pearson")

}
\references{
Wei T, Simko V (2021). R package 'corrplot': Visualization of a Correlation
Matrix. (Version 0.92), \url{https://github.com/taiyun/corrplot}.

Wickham H (2019). \emph{assertthat: Easy Pre and Post Assertions}. R package version 0.2.1,
\url{https://CRAN.R-project.org/package=assertthat}.
}
