% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{exprNormalization}
\alias{exprNormalization}
\title{Normalizes Gene Expression Data}
\usage{
exprNormalization(expressionData, method = "total")
}
\arguments{
\item{expressionData}{A dataframe of gene expression data, with genes as
rows and samples as columns.}

\item{method}{A parameter to specify normalization method. The options are:
\itemize{
\item "total" - total count normalization (default)
\item "log" - log2 transformation
\item "standard" - z-score standardization
}}
}
\value{
Returns a dataframe of of normalized gene expression data from
expressionData.
}
\description{
A function that calculates normalizes gene expression data, with three
methods available: total count normalization, log2 transformation, and
z-score standardization. The function uses the \code{dplyr} package to
manipulate the data.
}
\examples{
# Using OVExpression dataset in package

# Example 1:
# Using default normalization method "total"
normalizedExpr <- exprNormalization(OVExpression)
# view first 5 rows and columns
normalizedExpr[1:5, 1:5]

# sum of each column/sample is 1
colSums(normalizedExpr)


# Example 2:
# Using log2 transformation
logExpr <- exprNormalization(OVExpression, method = "log")
# view first 5 rows and columns
logExpr[1:5, 1:5]


# Example 3:
# Using standardization
standardizedExpr <- exprNormalization(OVExpression, method = "standard")
# view first 5 rows and columns
standardizedExpr[1:5, 1:5]

# mean for each column/sample is 0
round(colMeans(standardizedExpr))
# standard deviation for each column/sample is 1
apply(standardizedExpr,2,sd)

}
\references{
Wickham H, François R, Henry L, Müller K (2022). dplyr: A Grammar of Data
Manipulation. \url{https://dplyr.tidyverse.org},
\url{https://github.com/tidyverse/dplyr}.
}
