% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExpression.R
\name{exprPlot}
\alias{exprPlot}
\title{Plot Differential Gene Expression of Case vs Control}
\usage{
exprPlot(expressionData, sampleData, genes = NULL)
}
\arguments{
\item{expressionData}{A dataframe of numeric gene expression data, with genes as
rows and samples as columns.}

\item{sampleData}{A dataframe of sample information, with samples as rownames,
and a column called "type" identifying if the sample is case or control.}

\item{genes}{A string or character vector of genes to include in the plot.
This argument is optional; if not included, all genes will be included
in the plot. If input is not a string or character vector, all genes will
be included in the plot.}
}
\value{
Returns a boxplot of expression of specified genes in cases and controls.
}
\description{
Produces boxplots of gene expression data for specified genes. The function
uses the \code{melt} function from the \code{reshape} package to manipulate data, and
the \code{ggplot2} package to create the plot.
}
\examples{
# Using OVExpression and OVSample datasets in package

# Example 1:
# Exclude optional genes argument

# all genes included
exprPlot(expressionData = OVExpression,
        sampleData = OVSample)


# Example 2:
# Include genes argument
# Only genes PAX8 and DDR1 are included
exprPlot(expressionData = OVExpression,
        sampleData = OVSample,
        genes = c("PAX8", "DDR1"))


# Example 3:
# Using keggDiseaseGenes to get a list of genes associated with a disease
# Get genes associated with ovarian cancer (KEGG entry H00027)
OVgenes <- keggDiseaseGenes(entryID = "H00027")
exprPlot(expressionData = OVExpression,
        sampleData = OVSample,
        genes = OVgenes)

}
\references{
Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4, \url{https://ggplot2.tidyverse.org}.

Wickham H (2007). “Reshaping data with the reshape package.” Journal of
Statistical Software, 21(12). \url{https://www.jstatsoft.org/v21/i12/}.
}
