% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExpression.R
\name{exprPlot}
\alias{exprPlot}
\title{Plot Differential Gene Expression of Case vs Control}
\usage{
exprPlot(expressionData, sampleData, genes = NULL)
}
\arguments{
\item{expressionData}{A dataframe of gene expression data, with genes as
rows and samples as columns.}

\item{sampleData}{A dataframe of sample information, with two columns; one
for samples and a second identifying if the sample is case or control.}

\item{genes}{A string or character vector of genes to include in the plot.
This argument is optional; if not included, all genes will be included
in the plot. If input is not a string or character vector, all genes will
be included in the plot.}
}
\value{
Returns a dataframe with genes sorted
}
\description{
Produces boxplots of gene expression data for specified genes. The function
uses the \code{melt} function from the \code{reshape} package to manipulate data, and
the \code{ggplot2} package to create the plot.
}
\examples{
# Using OVExpression and OVSample datasets in package

# Example 1:
# Exclude optional genes argument

# all genes included
exprPlot(expressionData = OVExpression,
        sampleData = OVSample)


# Example 2:
# Include genes argument

# Only genes PAX8 and DDR1 are included
exprPlot(expressionData = OVExpression,
        sampleData = OVSample,
        genes = c("PAX8", "DDR1"))

}
\references{
Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4, \url{https://ggplot2.tidyverse.org}.

Wickham H (2007). “Reshaping data with the reshape package.” Journal of
Statistical Software, 21(12). \url{https://www.jstatsoft.org/v21/i12/}.
}
