% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnnotations.R
\name{geneAnnotations}
\alias{geneAnnotations}
\title{geneAnnotations (with AnnotationDbi).}
\usage{
geneAnnotations(input, keys, column, keytype)
}
\arguments{
\item{input}{Can be a) a data frame with keys as rownames or b) a vector of keys.}

\item{keys}{The keys to select records for from the database. Find keytypes with keytypes(org.Hs.eg.db).}

\item{column}{The columns or kinds of things that can be retrieved from the database. As with keys, all possible columns are returned by using columns(org.Hs.eg.db).}

\item{keytype}{Allows the user to discover which keytypes can be passed in to select or keys and the keytype argument.}
}
\value{
a) will return the same data frame as input with additional column(s) containing the annotation(s). b) will convert input to data rame and add column(s) containing the annotation(s).
}
\description{
Adds gene annotations
}
\examples{
geneAnnotations(input=DEgenes_pw[1:10,], keys=row.names(DEgenes_pw[1:10,]),
                 column=c("ENTREZID", "ENSEMBL"), keytype="SYMBOL")
}

