% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load data from files (csv/tsv/excel??)}
\usage{
loadData(exprFilePath, sampleFilePath, type = "tsv")
}
\arguments{
\item{exprFilePath}{A string of the path to the file containing gene
expression data. The file should be a matrix with the first row as sample
names, the first column as gene names, and expression values in the rest
of the matrix.}

\item{sampleFilePath}{A string of the path to the file containing sample
information. The file should have the first column as sample names, and
the second column as their type.}

\item{type}{A parameter to specify the file type. The options are:
\itemize{
\item "tsv" - tab-separated values (default)
\item "csv" - comma-separated values
}}
}
\value{
Returns a list of two dataframes:
\itemize{
\item "expressionData" - dataframe of expression values, rows are genes, columns are samples

\item "sampleData" - dataframe of sample information, rows are samples, one column "type" with the type of sample
}
}
\description{
A function prepares data in the format required for other functions in the
expressionAnalysis package, given the path to the files.
}
\examples{
# TODO!

}
\references{
Wickham H, Bryan J (2022). readxl: Read Excel Files.
https://readxl.tidyverse.org, https://github.com/tidyverse/readxl.
}
