% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_eigengene_plot.R
\name{module_eigengene_plot}
\alias{module_eigengene_plot}
\title{module_eigengene_plot (from WGCNA output).}
\usage{
module_eigengene_plot(groups, MEs, color)
}
\arguments{
\item{groups}{Define sample groups.}

\item{MEs}{Module eigengenes.}

\item{color}{Module color.}
}
\value{
Produces module eigengene plots.
}
\description{
This function implements module eigengene plots from wgcna output.
}
\examples{
module_eigengene_plot(groups, MEs, color)
}

