% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathview_func.R
\name{pathview_func}
\alias{pathview_func}
\title{pathview_func (from pathview functions).}
\usage{
pathview_func(DEgenes_pw, pathway.id, out.suffix)
}
\arguments{
\item{DEgenes_pw}{List of differentially expressed genes as output given by diff_limma_pairwise(). Rownames should be gene names (SMYBOLS).}

\item{pathway.id}{KEGG pathway identification number.}

\item{out.suffix}{Output suffix to be added to file name.}
}
\value{
Three output files are produced for each pathway: An .xml, a .png of the pathway and a .png of the pathway with DE genes and their fold change.
}
\description{
This produces pathview plots of differentially expressed genes and their fold changes for given pathways.
}
\details{
If you get the error message(s) data set 'korg' not found and/or data set 'bods' not found, it didn't properly extract these datasets from pathview package. Run data(korg) and data(bods) manually!
}
\examples{
pathview_func(DEgenes_pw, pathway.id = "04620", out.suffix = "DE_TolLPS")
                           #Toll-like-receptor signaling
}

