% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExpression.R
\name{rankDEG}
\alias{rankDEG}
\title{Rank Differentially Expressed Genes}
\usage{
rankDEG(expressionData, sampleData, case, control, method = "t")
}
\arguments{
\item{expressionData}{A dataframe of gene expression data, with genes as
rows and samples as columns.}

\item{sampleData}{A dataframe of sample information, with two columns; one
for samples and a second identifying if the sample is case or control.}

\item{case}{A string representing how samples are identified as cases in
sampleData.}

\item{control}{A string representing how samples are identified as controls
in sampleData.}

\item{method}{A parameter to specify the method used to calculate
differential expression. The possible parameters are:
\itemize{
\item "t" - t-test (default)
\item "wilcoxon" - Wilcoxon rank sum test
}}
}
\value{
Returns a dataframe with genes and P-values, in increasing order.
}
\description{
Identify and rank differentially expressed genes using expression data to
identify potential biomarkers.
}
\examples{
# Using OVExpression and OVSample datasets in package

# Example 1:
# Using default method t-test

genesTTest <- rankDEG(expressionData = OVExpression,
                      sampleData = OVSample,
                      case = "Ovarian cancer",
                      control = "Normal")
head(genesTTest)


# Example 2:
# Using Wilcoxon rank sum test
genesWilcoxon <- rankDEG(expressionData = OVExpression,
                      sampleData = OVSample,
                      case = "Ovarian cancer",
                      control = "Normal",
                      method = "wilcoxon")
head(genesWilcoxon)

}
